update Menu package

+++ src/main/java/org/maktab/View/Menu/AdminMenu.java
@@ -2,6 +2,7 @@ package org.maktab.View.Menu;
 
 import org.maktab.Check.Check;
 import org.maktab.Entity.Admin;
+import org.maktab.Entity.User;
 import org.maktab.Exception.SignUpSignInException;
 import org.maktab.Repository.Impl.AdminRepositoryImpl;
 import org.maktab.Service.AdminService;
@@ -10,9 +11,10 @@ import java.sql.SQLException;
 import java.util.Scanner;
 
 public class AdminMenu {
+    Scanner input = new Scanner(System.in);
     AdminService adminService = new AdminService(new AdminRepositoryImpl());
     Check check = new Check();
-    public Admin signUp(Scanner input) throws SQLException {
+    public Admin signUp() throws SQLException {
         System.out.print("Enter Your username : ");
         String username = input.next();
         System.out.print("Enter Your password : ");
@@ -26,7 +28,7 @@ public class AdminMenu {
         }
     }
 
-    public Admin signIn(Scanner input) throws SQLException {
+    public Admin signIn() throws SQLException {
         System.out.print("Enter Your username : ");
         String username = input.next();
         System.out.print("Enter Your password : ");
@@ -39,18 +41,30 @@ public class AdminMenu {
         }
     }
 
-    public Admin edit(Admin admin, Scanner input) throws SQLException {
+    public Admin edit(Admin admin) throws SQLException {
         System.out.print("Enter New username : ");
         String username = input.next();
         System.out.print("Enter New password : ");
         String password = input.next();
-        Admin admin1 = adminService.read(new Admin(username,password));
-        adminService.update(admin1,admin.getId());
-        return adminService.read(admin1);
+        Admin newAdmin = new Admin(username,password);
+        if (adminService.readByUsername(newAdmin)){
+            throw new SignUpSignInException("This username is Exist!");
+        }else {
+            adminService.update(newAdmin, admin.getId());
+            System.out.println("Edit Is Successfully.");
+            return adminService.read(newAdmin);
+        }
     }
 
     public void remove(Admin admin) throws SQLException {
-        adminService.delete(admin);
-        System.exit(1);
+        System.out.print("Are You Sure About Delete Your Account (Y/N) : ");
+        boolean yn = check.checkYN();
+        if (yn) {
+            adminService.delete(admin);
+            System.out.println("Delete Is Successfully.");
+            System.exit(1);
+        }else {
+            System.out.println("Delete Isn't Successfully.");
+        }
     }
 }
